+++ NEWS.md
@@ -3,6 +3,7 @@
 * `loadMSLibrary()`: improve compatibility with more `.msp` flavors (issue #72).
 * `newProject()`: save/load parameters to reproduce subsequent project creations (issue #61)
 * Fixed: in rare cases EICs were incorrectly loaded from cache
+* Fixed: `report()` now correctly handles SIRIUS compounds results and suspects without SMILES
 
 
 # patRoon 2.2